+++ src/syscall/syscalls/sys_exec_prog.cpp
@@ -4,7 +4,7 @@
  */
 #include "sys_exec_prog.h"
 
-sys_exec_prog::sys_exec_prog(size_t pid, program *prog)
+sys_exec_prog::sys_exec_prog(int pid, program *prog)
 	: syscall_t(syscall_id_t::EXEC_PROGRAM)
 {
 	this->pid = pid;
